Long long batches

+++ src/bugger/modules/SideModul.hpp
@@ -53,7 +53,7 @@ namespace bugger {
          if( solution.status == SolutionStatus::kUnbounded )
             return ModulStatus::kNotAdmissible;
 
-         int batchsize = 1;
+         long long batchsize = 1;
 
          if( parameters.nbatches > 0 )
          {
